Provenance Attributes
These attributes can occur in netCDF enhanced (netcdf-4) files beginning with version 4.4.1. They all are associated with the root group as global attributes, although only _NCProperties is actually stored in the file; the others are computed.

if attribute name is '_NCProperties', transforme to 'NCProperties_'
because Matlab doesn't handle member name begining with '_'